fix tests and lint
