+++ packages/ignite-unholy-app-template/templates/App/Services/Api.js
@@ -43,7 +43,7 @@ const create = (baseURL = 'https://api.github.com/') => {
   //
   const getRoot = () => api.get('')
   const getRate = () => api.get('rate_limit')
-  const getUser = (username) => api.get('search/users', {q: username})  
+  const getUser = (username) => api.get('search/users', {q: username})
 
   // ------
   // STEP 3
