+++ packages/ignite-unholy-app-template/templates/Tests/Sagas/StartupSagaTest.js
@@ -1,13 +1,13 @@
 import test from 'ava'
 import { select, put } from 'redux-saga/effects'
-import { selectTemperature, startup } from '../../App/Sagas/StartupSagas'
-import TemperatureActions from '../../App/Redux/TemperatureRedux'
+import { selectAvatar, startup } from '../../App/Sagas/StartupSagas'
+import GithubActions from '../../App/Redux/GithubRedux'
 
 const stepper = (fn) => (mock) => fn.next(mock).value
 
 test('watches for the right action', (t) => {
   const step = stepper(startup())
-  TemperatureActions.temperatureRequest('San Francisco')
-  t.deepEqual(step(), select(selectTemperature))
-  t.deepEqual(step(), put(TemperatureActions.temperatureRequest('San Francisco')))
+  GithubActions.userRequest('GantMan')
+  t.deepEqual(step(), select(selectAvatar))
+  t.deepEqual(step(), put(GithubActions.userRequest('GantMan')))
 })
